// src/models/vote-info.ts
export class VoteInfo{

      gameNo: Number;
      personId: Number;
      tournamentId: Number;
      teamAId: Number;
      teamBId : Number;
      teamAValue: Number;
      teamBValue: Number;
      participated: Boolean;
      //voteingTime:Date;

			message:String;
			modifiedcounter:Number;
			updatedStatus:Boolean;
}